Ta ut rendere riktig oppsummering seksjon og versjon til egen komponent og fjern html break lines
